Email-marketer/requirements.txt
fastapi
uvicorn
motor
pydantic
pydantic-settings
python-dotenv
email-validator
anthropic
gspread
oauth2client
langgraph
langchain-anthropic
langchain
langchain-core
google-auth
google-auth-oauthlib
google-api-python-client
cryptography
python-jose[cryptography]
dnspython
watchfiles
